requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
requests
<<<<<<< HEAD
twilio-rest
weather-checker
=======
twilio
>>>>>>> 6eec79cd7bb8e8e3d9aa0c2e72840329bfb5a740
